% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY_area_source_emissions.R, R/deprecated.R
\name{BY_area_source_emissions}
\alias{BY_area_source_emissions}
\alias{area_source_emissions}
\title{BY_area_source_emissions}
\usage{
BY_area_source_emissions(base_year, ..., verbose = getOption("verbose"))
}
\description{
BY_area_source_emissions
}
