% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BY_area_source_projections.R, R/deprecated.R
\name{BY_area_source_projections}
\alias{BY_area_source_projections}
\alias{BY_area_source_projections_}
\alias{area_source_projections}
\title{Calculate, Forecast, and Backcast Emissions for Area Sources}
\usage{
BY_area_source_projections(
  base_year = NULL,
  years = CY(1990:2040),
  tput_data = NULL,
  ef_data = NULL,
  cf_data = NULL,
  gf_data = NULL,
  na.rm = TRUE,
  verbose = getOption("verbose")
)
}
\arguments{
\item{base_year}{`BY` object, like `BY(2011)`}

\item{years}{`CY` object, like `CY(1990:2050)`; the range of years you want results for}

\item{tput_data}{see Details, below}

\item{ef_data}{see Details, below}

\item{cf_data}{see Details, below}

\item{na.rm}{passed to subroutines; refer to "See Also", below}

\item{verbose}{display informative messages}

\item{gpf_data}{see Details, below}
}
\value{
Tabular data:
- keyed by year, county, pollutant, and category;
- having columns `tput_qty`, `ef_qty`, `cf_qty`, `gf_qty`, and their product, `ems_qty`.
}
\description{
`BY_area_source_projections()` combines throughputs, emission factors, control
factors, and growth profiles, multiplying them together. This results in a
set of "grown and controlled" emissions for the given `years`.
}
\details{
**BASIC USAGE**

If you don't supply any or all of:

  - `tput_data`
  - `ef_data`
  - `cf_data`
  - `gpf_data`

... then, helpfully, `BY_area_source_projections()` will fill them in for you.
In that case, so that it knows where to look, you will have to supply a base
year, like so:

```
    BY(2011) %>% BY_area_source_projections()
```

In this case, it's helpful to pass `verbose = TRUE`, so that you'll know
what's happening and where the throughput, emission factor, control factor,
and growth profile data are being pulled from.

**ADVANCED USAGE**

You can, instead, BYOD ("bring your own data"). Then you can use this function to
explore combinations of alternate throughputs, growth factors, emission
factors, and control factors.

You can simply read in a CSV file, or an XLSX worksheet, containing the
`tput_data`, `ef_data`, `cf_data`, or `gpf_data` that you would like to use.
(You will have to use `read_csv()` or `read_excel()` to import the data
first.)

The data that you do supply should look like the defaults:

- `tput_data` defaults to `DB_area_source_throughputs(base_year)`.
- `ef_data` defaults to `annualize_emission_factors(DB_area_source_emission_factors(base_year))`.
- `cf_data` defaults to `annualize_control_factors(DB_area_source_control_factors(base_year))`.
- `gpf_data` defaults to `DB_growth_profiles(base_year))`.

Try out those defaults, or look at the help pages. (Click the links under "See Also", below.)
As long as you supply similar-looking data, you are good to go.
}
\note{
If you don't supply `tput_data`, then data will be fetched, and results
  will be calculated, for every single category in the given base year. If
  you do supply `tput_data`, the result will be scoped to just the categories
  present in `tput_data$cat_id`. (This will be noticeably faster if the number of categories is small.)
}
\seealso{
- [DB_area_source_throughputs()]
- [DB_area_source_emission_factors()] and [annualize_emission_factors()]
- [DB_area_source_control_factors()] and [annualize_control_factors()]
- [DB_growth_profiles()]
}
