% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_counties.R
\name{expand_counties}
\alias{expand_counties}
\title{Expand all cases where `cnty_abbr` is either "TOT" or `NA`,
replacing with the full set of nine county codes.}
\usage{
expand_counties(
  input_data,
  cnty_var = "cnty_abbr",
  cnty_codes = NULL,
  verbose = getOption("verbose")
)
}
\description{
Expand all cases where `cnty_abbr` is either "TOT" or `NA`,
replacing with the full set of nine county codes.
}
