% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_annual_throughputs_by.R
\name{project_annual_throughputs_by}
\alias{project_annual_throughputs_by}
\title{project_annual_throughputs_by}
\usage{
project_annual_throughputs_by(..., verbose = getOption("verbose"))
}
\description{
project_annual_throughputs_by
}
